Separate unit from integration tests
