+++ tests/integration/test_infer.py
