+++ tests/units/training/test_train_config.py
@@ -7,9 +7,7 @@ import os
 
 import pytest
 import tomli
-from pydantic import ValidationError
 
-from zeroband.infer import Config as InferenceConfig
 from zeroband.train import Config as TrainingConfig
 
 
@@ -28,17 +26,3 @@ def test_load_train_configs(config_file_path):
         content = tomli.load(f)
     config = TrainingConfig(**content)
     assert config is not None
-
-
-@pytest.mark.parametrize("config_file_path", get_all_toml_files("configs/inference"))
-def test_load_inference_configs(config_file_path):
-    with open(f"{config_file_path}", "rb") as f:
-        content = tomli.load(f)
-    config = InferenceConfig(**content)
-    assert config is not None
-
-
-def test_throw_error_for_dp_and_pp():
-    with pytest.raises(ValidationError):
-        config = InferenceConfig(**{"dp": 2, "pp": {"world_size": 2}})
-        print(config)
